+++ DataDefinitions/LegalStatus.cs
@@ -12,17 +12,25 @@ namespace EddiDataDefinitions
         {
             resourceManager = Properties.LegalStatus.ResourceManager;
             resourceManager.IgnoreCase = false;
+            missingEDNameHandler = (edname) => new LegalStatus(edname);
 
-            var None = new LegalStatus("None");
+            Clean = new LegalStatus("Clean");
+            None = new LegalStatus("None");
             var Unknown = new LegalStatus("Unknown");
             var Lawless = new LegalStatus("Lawless");
-            var Clean = new LegalStatus("Clean");
             var Wanted = new LegalStatus("Wanted");
             var Enemy = new LegalStatus("Enemy");
             var WantedEnemy = new LegalStatus("WantedEnemy");
             var Warrant = new LegalStatus("Warrant");
             var Hunter = new LegalStatus("Hunter");
+            var IllegalCargo = new LegalStatus("IllegalCargo");
+            var Speeding = new LegalStatus("Speeding");
+            var Hostile = new LegalStatus("Hostile");
+            var PassengerWanted = new LegalStatus("PassengerWanted");
         }
+
+        public static readonly LegalStatus Clean;
+        public static readonly LegalStatus None;
 
         // dummy used to ensure that the static constructor has run
         public LegalStatus() : this("")